requirements.txt
fastapi
uvicorn
PyPDF2
python-docx
chromadb
scikit-learn
PyYAML

